Climatic data from the Greek National Meteorological Service

The climatic data are taken from the Greek National Meteorological Service and come from 21 weather stations, with a time step of 3 hours. The time period is from 1975 to 2004. These time series were not checked for consistency or errors.

emy_data

Format

A data frame with 1679308 rows and 13 variables:

date

The observation's date

station_id

Greek National Meteorological Service station ID

pressure

Pressure, in hPa

dry_temp

Dry temperature, in degree Celsius

wet_temp

Wet temperature, in degree Celsius

max_temp

Maximum temperature, in degree Celsius

min_temp

Minimum temperature, in degree Celsius

rel_humidity

Relative humidity, in percent

wind_direction

Wind direction, in degrees

wind_force_knot

Wind force, in knots

visibility

Visibility, in km

total_cloud_amount

Total cloud amount, in values 0 to 9

precipitation

Precipitation height, in mm

Source

http://www.emy.gr/