emyweather: Open data from the Greek National Meteorological Service

This package contains climatic data from the Greek National Meteorological Service, as well as a number of useful other related data sets.

Climatic Data

The climatic data are taken from the Greek National Meteorological Service and come from 21 weather stations, with a time step of 3 hours. The time period is from 1975 to 2004. The variables are: air pressure, temperature (dry, wet, max and min), relative humidity, wind (direction and speed), visibility, total cloud amount and precipitation. These time series were not checked for consistency or errors.

Stations

The stations' data are taken from the Greek National Data Bank for Hydrological and Meteorological Information, Hydroscope. The variables are: stations' id, coordinates, altitude and water division.

Data Sources

The data are retrieved from: http://www.meteo.gr/ http://www.hydroscope.gr/

See also

Useful links:

https://github.com/kvantas/emyweather

Report bugs at https://github.com/kvantas/emyweather/issues